Avoid BC-break with symfony/form 5
