+++ src/Form/Extension/Field/Type/MopaCompatibilityTypeFieldExtension.php
@@ -13,45 +13,41 @@ declare(strict_types=1);
 
 namespace Sonata\AdminBundle\Form\Extension\Field\Type;
 
-use Symfony\Component\Form\AbstractTypeExtension;
 use Symfony\Component\Form\Extension\Core\Type\FormType;
-use Symfony\Component\Form\FormInterface;
-use Symfony\Component\Form\FormView;
-use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Form\FormTypeExtensionInterface;
 
-/**
- * This class is built to allow AdminInterface to work properly
- * if the MopaBootstrapBundle is not installed.
- *
- * @final since sonata-project/admin-bundle 3.52
- *
- * @author Thomas Rabaix <thomas.rabaix@sonata-project.org>
- */
-class MopaCompatibilityTypeFieldExtension extends AbstractTypeExtension
-{
-    public function configureOptions(OptionsResolver $resolver)
+// NEXT_MAJOR: Remove the "else" part, copy all methods from BaseMopaCompatibilityTypeFieldExtension in this class and
+// extend from AbstractTypeExtension
+if (method_exists(FormTypeExtensionInterface::class, 'getExtendedTypes')) {
+    /**
+     * This class is built to allow AdminInterface to work properly
+     * if the MopaBootstrapBundle is not installed.
+     *
+     * @final since sonata-project/admin-bundle 3.52
+     *
+     * @author Thomas Rabaix <thomas.rabaix@sonata-project.org>
+     */
+    class MopaCompatibilityTypeFieldExtension extends BaseMopaCompatibilityTypeFieldExtension
     {
-        $resolver->setDefaults([
-            'horizontal_label_class' => '',
-            'horizontal_label_offset_class' => '',
-            'horizontal_input_wrapper_class' => '',
-        ]);
+        public static function getExtendedTypes(): iterable
+        {
+            return [FormType::class];
+        }
     }
-
-    public function buildView(FormView $view, FormInterface $form, array $options)
+} else {
+    /**
+     * This class is built to allow AdminInterface to work properly
+     * if the MopaBootstrapBundle is not installed.
+     *
+     * @final since sonata-project/admin-bundle 3.52
+     *
+     * @author Thomas Rabaix <thomas.rabaix@sonata-project.org>
+     */
+    class MopaCompatibilityTypeFieldExtension extends BaseMopaCompatibilityTypeFieldExtension
     {
-        $view->vars['horizontal_label_class'] = $options['horizontal_label_class'];
-        $view->vars['horizontal_label_offset_class'] = $options['horizontal_label_offset_class'];
-        $view->vars['horizontal_input_wrapper_class'] = $options['horizontal_input_wrapper_class'];
-    }
-
-    public function getExtendedType()
-    {
-        return FormType::class;
-    }
-
-    public static function getExtendedTypes(): iterable
-    {
-        return [FormType::class];
+        public static function getExtendedTypes()
+        {
+            return [FormType::class];
+        }
     }
 }
